+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-01-16
+
+### Added
+- Auto Restart Cursor Process
+- Auto Reset Machine Ids
+
+### Fixed
+- Fix some known issues
+
 ## [1.0.9] - 2024-01-15
 
 ### Fixed